ai_agents/call_performance_analyzer/requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Pydantic for schemas
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0

